pyproject.toml
[project]
name = "llm-project-12-22-25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langfuse>=3.11",
    "openai>=2.14.0",
    "pinecone>=8.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "yagmail>=0.15.293",
]
